Change Config signatures to avoid clones
It seems like Config uses the internal Arc to know if it has already
been initialized or not.

When calling it from FFI on Dart, I had seem the following warnings:

```
config has already been used to start the system and probably should not be mutated
```

This commit removes the intermediate clones. It now expects the config
to be passed the ownership back to Rust, and return a new Config, just
like how it does on Rust.

Beware of danling pointers from the C/FFI side.