Fix stacked borrows violation in CacheAccessIter
The pointer `self.current_block` is separately turned into a shared
reference and a Box in the same method, and moreover, the reference is
created before the Box and used after the Box. This results in a unique
reference from the Box and a shared reference from a separate
pointer-to-reference conversion coexisting, with different tags.

To fix this issue, the conversion from a raw pointer to a Box is put off
until the shared reference is no longer in use, so that the two
different tags won't pop each other off the stack.